products/requirements.txt
fastapi
uvicorn
motor            # Driver assíncrono para MongoDB
pydantic
pydantic-settings
python-dotenv
structlog        # O "Pino" do Python
email-validator
pyjwt